Fixed elapsed time not accurately showing time

+++ src/util.c
@@ -8,6 +8,6 @@ int rand_range(const int min, const int max) {
 double calculate_elapsed(struct timeval* start, struct timeval* end) {
     long seconds = end->tv_sec - start->tv_sec;
     long microseconds = end->tv_usec - start->tv_usec;
-    double time_elapsed = seconds + (microseconds * MICRO_LENGTH);
+    double time_elapsed = seconds + ((double)microseconds * MICRO_LENGTH);
     return time_elapsed;
 }
